More naming consistency, folded useless scripts
